+++ metrics/src/main/scala/org/broadinstitute/dsde/workbench/metrics/WorkbenchInstrumented.scala
@@ -2,7 +2,8 @@ package org.broadinstitute.dsde.workbench.metrics
 
 import akka.http.scaladsl.model.{HttpRequest, HttpResponse}
 import com.codahale.metrics.{Gauge => DropwizardGauge}
-import nl.grons.metrics.scala._
+import nl.grons.metrics.scala.{DefaultInstrumented, MetricName}
+import nl.grons.metrics.scala.{Gauge => GronsGauge}
 import org.broadinstitute.dsde.workbench.metrics.Expansion._
 import scala.collection.JavaConverters._
 
@@ -18,6 +19,8 @@ trait WorkbenchInstrumented extends DefaultInstrumented {
     */
   protected val workbenchMetricBaseName: String
   override lazy val metricBaseName = MetricName(workbenchMetricBaseName)
+
+  implicit def metricUnwrapper[M](metric: Metric[M]): M = metric.metric
 
   /**
     * Utility for building expanded metric names in a typesafe way. Example usage:
@@ -41,15 +44,22 @@ trait WorkbenchInstrumented extends DefaultInstrumented {
         (if (m == "") m else m + ".") + implicitly[Expansion[A]].makeNameWithKey(key, a))
     }
 
-    def asCounter(name: String): Counter =
-      metrics.counter(makeName(name))
+    def getFullName(name: String): String =
+      metricBaseName.append(makeName(name)).name
 
-    def asGauge[T](name: String)(fn: => T): Gauge[T] =
-      metrics.gauge(makeName(name))(fn)
+    def asCounter(name: String): Counter = {
+      val cnt = metrics.counter(makeName(name))
+      new Counter(getFullName(name), cnt)
+    }
+
+    def asGauge[T](name: String)(fn: => T): Gauge[T] = {
+      val gauge = metrics.gauge(makeName(name))(fn)
+      new Gauge[T](getFullName(name), gauge)
+    }
 
     def asGaugeIfAbsent[T](name: String)(fn: => T): Gauge[T] = {
       // Get the fully qualified metric name for inspecting the registry.
-      val gaugeName = metricBaseName.append(makeName(name)).name
+      val gaugeName = getFullName(name)
       metricRegistry.getGauges().asScala.get(gaugeName) match {
         case None =>
           // If the gauge does not exist in the registry, create it
@@ -57,15 +67,23 @@ trait WorkbenchInstrumented extends DefaultInstrumented {
         case Some(gauge) =>
           // If the gauge exists in the registry, return it.
           // Need to wrap the returned Java DropwizardGauge in a Scala Gauge.
-          new Gauge[T](gauge.asInstanceOf[DropwizardGauge[T]])
+          new Gauge[T](gaugeName, new GronsGauge[T](gauge.asInstanceOf[DropwizardGauge[T]]))
       }
     }
 
-    def asTimer(name: String): Timer =
-      metrics.timer(makeName(name))
+    def asTimer(name: String): Timer = {
+      val tim = metrics.timer(makeName(name))
+      new Timer(getFullName(name), tim)
+    }
 
-    def asHistogram(name: String): Histogram =
-      metrics.histogram(makeName(name))
+    def asHistogram(name: String): Histogram = {
+      val histo = metrics.histogram(makeName(name))
+      new Histogram(getFullName(name), histo)
+    }
+
+    def unregisterMetric[T](metric: Metric[T]): Boolean = {
+      metricRegistry.remove(metric.name)
+    }
 
     private def makeName(name: String): String =
       if (m.nonEmpty) s"$m.$name" else name